Convert perf_counters::counter_type enum to enum class.

+++ libs/full/performance_counters/src/counters.cpp
@@ -506,7 +506,7 @@ namespace hpx { namespace performance_counters {
     {
         if (type < counter_text || type > counter_raw_values)
             return "unknown";
-        return strings::counter_type_names[type];
+        return strings::counter_type_names[static_cast<int>(type)];
     }
 
     ///////////////////////////////////////////////////////////////////////////
